Give Xed a Sublime-like searchbar
The search dialog is replaced with a searchbar similar to the one used in Sublime.

The following features are implicit and therefore removed:
  - "incremental search"
  - "remove highlights"
  - "parse \n characters"

Code style is applied to impacted files (should have been done in a separate commit but it was being discussed during WIP and the new committed code followed neither the new style nor the old).