toolbar: Revamp the main toolbar style
Get rid of the text and switch all icons to slightly smaller symoblic icons.
This will allow it to behave much better with dark gtk themes. Replace the
recent documents button with a button for file open. Changing the style of
the buttons on the main toolbar also effects the fullscreen toolbar so
update this as well to use a matching layout.