Implement font zooming in the editor
Give controls for zoom in, zoom out, and reset zoom in the menu and via hotkey.
Also add a new placeholder in the view menu to allow positioning the menuitems
properly. Still want to keep the highlighting menu item at the  bottom.

This is implemented by using a builtin libpeas python plugin. Add support for
loading these plugins at startup. They cannot be enabled or disabled through the
UI. They can still be disabled in gsettings but will be automatically reloaded
when xed is restarted.

Closes: https://github.com/linuxmint/xed/issues/45