Minimap (#175)
* Add a minimap

Add support for GtkSourceViews SourceMap object. Add both a default setting for
new files to the preferences dialog and a menu item to allow temporarily
disabling it in any view.

* Require gtk and gtksourceview 3.18

* load custom css

We can use this to make small tweaks to the interface so every theme doesn't
need to add specific support.

Make use of this new css to clean up the styling of the new overview map frame
a bit.