search bar/xed-history-entry: Set a 'width-chars' on the GtkEntry widgets of the search and replace entry widgets, and allow their labels to ellipsize upon resizing.
This allows the entry widgets to have a smaller minimum size, which
was having an impact on window tiling for certain locales and display
resolutions (muffin and other wm's won't allow tiling of a window
if the resulting window size is less than the minimum size hints
reported by gtk and other toolkits.)

Ellipsizing the labels has the same benefit of enabling xed to be
resized smaller to tile on smaller displays.  This ellipsizing only
occurs once the entries hit their minimum size.

Fixes #73