Fix stuttering and disappearing of the toolbar when fullscreen (#300)
This fixes an issue where moving the mouse along the top edge of the
screen while fullscreened will cause a leave event to be triggered,
even though the mouse pointer is still inside the event box. This
caused the toolbar to disappear and/or stutter while moving the mouse
along the top edge.