Revert "dark mode: Delegate to libxapp"
This reverts commit 233bc4afed114675376534182b8c83d1d6c6d193.

Xed is a bit of an exception. It's light by default but many people
want to be able to make it dark.

Using prefer-dark globally with a light theme is a no-go, it makes
firefox and many other apps dark.

This revert brings back the ability to make Xed dark without using
the global setting.